requirements.txt
streamlit>=1.36.0
Pillow>=10.0.0
piexif>=1.1.3
geopy>=2.4.1
